requirements.txt
requests>=2.31.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0
